+++ src/local-development/code-pull-deploy.ts
@@ -131,6 +131,16 @@ export async function deployComponentCode({
 
 	const codeDef = getCodeDef(appComponentType, codeType);
 
+	try {
+		await vscode.workspace.fs.stat(sourcePath);
+	} catch (e: any) {
+		if (e.code === 'FileNotFound') {
+			throw new Error('Skipped code deployment, because local file is missing.');
+		} else {
+			// Unknown error
+			throw e;
+		}
+	}
 	const sourceContentUint8 = await vscode.workspace.fs.readFile(sourcePath);
 	const sourceContent = new TextDecoder().decode(sourceContentUint8);